+++ gradsflow/__init__.py
@@ -1,8 +1,8 @@
 """An AutoML Library made with Optuna and PyTorch Lightning"""
 
-from gradsflow.tasks.autoclassification.text import AutoTextClassifier
-from gradsflow.tasks.autoclassification.image import AutoImageClassifier
-from gradsflow.tasks.summarization import AutoSummarization
+from gradsflow.autotasks.autoclassification.text import AutoTextClassifier
+from gradsflow.autotasks.autoclassification.image import AutoImageClassifier
+from gradsflow.autotasks.autosummarization import AutoSummarization
 from gradsflow.core.automodel import AutoModel
 
 __version__ = "0.0.1"
