+++ tests/core/test_automodel.py
